# engines/contracts/engine-v1.openapi.yaml
openapi: 3.1.0
info:
  title: Curf Engine API
  version: 1.0.0
  summary: Data plane for Curf - connections, governed views, queries, policy and audit.
  description: |
    Contract for the Curf engine. Every implementation must pass the shared conformance suite in
    `engines/conformance`. Errors are RFC 9457 `application/problem+json` with a stable `code`.
    Milestone M0 covers identity, audit and operations; later milestones add connections, views,
    queries, reports, exports, sharing and schedules.
  license:
    name: MIT
    identifier: MIT
  contact:
    name: Curf engine maintainers
    url: https://github.com/EnterpriseX-Platform/CurfAI
servers:
  - url: http://localhost:8080/engine/v1
    description: Local engine
...
security:
  - bearerAuth: []
paths:
...
  /reports:
    get:
      operationId: listReports
      tags: [Reports]
      summary: List reports
      description: |
        Holders of `report:edit` get a page of every report. Everyone else gets the reports their roles may run.
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 0, default: 0 } }
        - { name: size, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
      responses:
        "200":
          description: A page of summaries (editors) or an array of runnable reports (everyone else).
          content:
            application/json:
              schema:
                oneOf:
                  - { $ref: "#/components/schemas/ReportPage" }
                  - { type: array, items: { $ref: "#/components/schemas/RunnableReport" } }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "422": { $ref: "#/components/responses/InvalidInput" }
    post:
      operationId: createReport
      tags: [Reports]
      summary: Save a report
      description: |
        Requires `report:edit`. The body carries a report in Curf's own definition format (see `ReportDefinition`), which
        is stored exactly as sent, as version 1.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReportRequest" }
      responses:
        "201":
          description: Saved.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Report" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/InvalidInput" }
...
  /schedules/{id}:
    get:
      operationId: getSchedule
      tags: [Schedules]
      summary: One schedule
      description: Needs `schedule:manage`.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: The schedule.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Schedule" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      operationId: updateSchedule
      tags: [Schedules]
      summary: Change a schedule
      description: |
        Saving runs the schedule as the person saving from now on (their rows, their masks); nobody can make a schedule
        run with more access than they hold. The report cannot be changed. `enabled: false` switches it off.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ScheduleRequest" }
      responses:
        "200":
          description: The schedule.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Schedule" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/InvalidInput" }
    delete:
      operationId: deleteSchedule
      tags: [Schedules]
      summary: Delete a schedule and its run history
      description: Needs `schedule:manage`.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Deleted. }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthenticated:
      description: Missing, invalid or untrusted token.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Forbidden:
      description: The caller lacks the required permission.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    InvalidInput:
      description: The request is malformed or fails validation. For SQL the code is `CURF_SQL_REJECTED` and `errors` says why.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: No such resource in the caller's tenant.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string, examples: ["urn:curf:error:forbidden"] }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code:
          type: string
          enum:
            - CURF_INVALID_INPUT
            - CURF_SQL_REJECTED
            - CURF_QUERY_TIMEOUT
            - CURF_TOO_MANY_QUERIES
            - CURF_UNAUTHENTICATED
            - CURF_FORBIDDEN
            - CURF_SEGREGATION_OF_DUTIES
            - CURF_NOT_FOUND
            - CURF_STALE_VERSION
            - CURF_CONFLICT
            - CURF_ROW_LIMIT
            - CURF_EXPORT_UNAVAILABLE
            - CURF_CONNECTION_FAILED
            - CURF_INTERNAL
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
...
    ReportDefinition:
      type: object
      required: [version, name, pages]
      description: |
        Curf's report definition, stored as sent (pages, blocks, themes and the rest are presentation the engine does not
        read). The engine reads `parameters` and `dataSources`, and which queries blocks use (any `queryId` or `*QueryId`
        value; `config.drilldown.queryId` marks drill-only queries). Each `dataSources[]` entry needs `id`, `name`,
        `dataSourceId` and exactly one of:

        - `engine`: `{viewId, columns?, filters?, groupBy?, aggregates?, orderBy?, limit?, params?}`, a published view plus
          structured parts, run as the caller. A value may be `{"$param": "name"}` to use a report parameter; a filter with
          `skipIfEmpty` is left out when its parameter is empty (Curf's "blank means no filter").
        - `sql`: a raw SELECT with `:name` parameters on the engine connection `dataSourceId`. Only for `view:manage` holders
          on connections that allow raw SQL; for everyone else the query is marked `accessDeniedNote`.

        Not supported yet: REST fields, `joins` and `attaches`.
      properties:
        version: { type: integer, enum: [1] }
        name: { type: string, minLength: 1, maxLength: 200 }
        parameters:
          type: array
          items:
            type: object
            required: [name, type]
            properties:
              name: { type: string, pattern: "^[A-Za-z_][A-Za-z0-9_]{0,63}$", description: Must not start with `__`. }
              label: { type: string }
              type: { type: string, enum: [string, number, date, dateRange, boolean, select] }
              required: { type: boolean }
              default: { type: [string, number, boolean] }
              options:
                type: array
                items:
                  type: object
                  properties:
                    value: { type: string }
                    label: { type: string }
        dataSources: { type: array, maxItems: 100, items: { type: object } }
        pages: { type: array, minItems: 1, items: { type: object } }
      additionalProperties: true
    ReportRequest:
      type: object
      required: [definition]
      properties:
        definition: { $ref: "#/components/schemas/ReportDefinition" }
        runRoles: { type: array, items: { type: string }, description: Roles that may run the report. Editors always may. }
        version: { type: integer, description: Required on update. }
        note: { type: string, maxLength: 500, description: Describes the change in the history. }
    Report:
      type: object
      required: [id, name, definition, runRoles, version]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        definition: { $ref: "#/components/schemas/ReportDefinition" }
        runRoles: { type: array, items: { type: string } }
        version: { type: integer }
        createdAt: { type: string, format: date-time }
        createdBy: { type: string }
        updatedAt: { type: string, format: date-time }
        updatedBy: { type: string }
    ReportSummary:
      type: object
      required: [id, name, version]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        runRoles: { type: array, items: { type: string } }
        version: { type: integer }
        updatedAt: { type: string, format: date-time }
        updatedBy: { type: string }
    ReportPage:
      type: object
      required: [content, page, size, totalElements]
      properties:
        content: { type: array, items: { $ref: "#/components/schemas/ReportSummary" } }
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer, format: int64 }
    RunnableReport:
      type: object
      required: [id, name, version]
      description: A report as someone who may only run it sees it. Queries are reduced to `id` and `name` when a definition is included.
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        version: { type: integer }
        definition: { $ref: "#/components/schemas/ReportDefinition" }
...
    ScheduleRequest:
      type: object
      required: [reportId, name, cron, format, recipients]
      properties:
        reportId: { type: string, format: uuid }
        name: { type: string, maxLength: 200 }
        cron: { type: string, description: "Five fields: minute hour day-of-month month day-of-week.", examples: ["30 6 * * 1-5"] }
        timezone: { type: string, description: IANA zone, default: UTC, examples: [Asia/Bangkok] }
        format: { type: string, enum: [CSV, XLSX, DOCX, PDF] }
        locale: { type: string, enum: [th, en] }
        params: { type: object, description: Report parameter values for every run. }
        recipients: { type: array, maxItems: 20, items: { type: string, format: email } }
        enabled: { type: boolean, default: true }
    Schedule:
      type: object
      properties:
        id: { type: string, format: uuid }
        reportId: { type: string, format: uuid }
        name: { type: string }
        cron: { type: string }
        timezone: { type: string }
        format: { type: string, enum: [CSV, XLSX, DOCX, PDF] }
        locale: { type: [string, "null"] }
        params: { type: object }
        recipients: { type: array, items: { type: string } }
        runAs: { type: string, description: Subject of the person it runs as. }
        enabled: { type: boolean }
        nextRunAt: { type: [string, "null"], format: date-time }
        createdAt: { type: string, format: date-time }
        createdBy: { type: string }
        updatedAt: { type: string, format: date-time }
        updatedBy: { type: string }
...